feat(auth): Implement user authentication
- AuthService is responsible for the register, login, and refresh operation.
Its methods accept and/or return request and response DTOs.
AuthService uses the refresh duration from the application.properties
for the refresh tokens. The refresh duration can also be set from the
.env file as JWT_REFRESH_DAYS
When it refreshes access and return new JTW tokens, it also rotates
the refresh token. This means that it invalidates the used refresh,
token and issues a new one. This is for security purposes as it limits
exploits through compromised tokens.

- AuthController maps the endpoints for authentication and to refresh
access tokens. Its methods accept request DTOs from the client,
and it uses the AuthService to generate the response DTOs, when
applicable.

- Created DTOs for Login, Refresh and Register Requests and
for Authentication Response.
The request DTOs ensure that the client only inputs the necessary
fields.
It also prevents over-posting meaning that unauthorized fields
cannot be modified through the client by modifying the request.
The response DTOs ensure that the backend outputs only what the
client needs. In this way, the sensitive or unecessary information
is not exposed.

- Isolated PasswordEncoder into it's own Config file and made it a @Bean
This way it can be @Autowired for auto dpendency injection.
The PasswordEncoder is now @Autowired in AuthProviderConfig.

- Changed the @GeneratedValue strategy to the default one (AUTO) for
the @Id property of the User entity. The 'identity' generation type
could only accept integers. The id of the User is of type UUID.

**Key Points:**
- DTOs are used to communicate between client and server.
- AuthService generates the response DTOs
- AuthController passes the request DTOs to AuthService
- Register, Login and Refresh operations work as intended

Closes #2, Closes #3, Closes #4